+++ lib/appium_lib_core/common/base/has_network_connection.rb
@@ -30,7 +30,7 @@ module Appium
         end
 
         def network_connection_type=(connection_type)
-          raise ArgumentError, 'Invalid connection type' unless valid_type? connection_type
+          raise ::Appium::Core::Error::ArgumentError, 'Invalid connection type' unless valid_type? connection_type
 
           connection_value = type_to_values[connection_type]