+++ lib/appium_lib_core/common/device/keyevent.rb
@@ -25,8 +25,8 @@ module Appium
           end
 
           def press_keycode(key, metastate: [], flags: [])
-            raise ArgumentError, 'flags should be Array' unless flags.is_a? Array
-            raise ArgumentError, 'metastates should be Array' unless metastate.is_a? Array
+            raise ::Appium::Core::Error::ArgumentError, 'flags should be Array' unless flags.is_a? Array
+            raise ::Appium::Core::Error::ArgumentError, 'metastates should be Array' unless metastate.is_a? Array
 
             args             = { keycode: key }
             args[:metastate] = metastate.reduce(0) { |acc, meta| acc | meta } unless metastate.empty?
@@ -36,8 +36,8 @@ module Appium
           end
 
           def long_press_keycode(key, metastate: [], flags: [])
-            raise ArgumentError, 'flags should be Array' unless flags.is_a? Array
-            raise ArgumentError, 'metastates should be Array' unless metastate.is_a? Array
+            raise ::Appium::Core::Error::ArgumentError, 'flags should be Array' unless flags.is_a? Array
+            raise ::Appium::Core::Error::ArgumentError, 'metastates should be Array' unless metastate.is_a? Array
 
             args             = { keycode: key }
             args[:metastate] = metastate.reduce(0) { |acc, meta| acc | meta } unless metastate.empty?