+++ test/unit/android/device/w3c/keyboard_test.rb
@@ -83,14 +83,14 @@ class AppiumLibCoreTest
 
           # keypress
           def test_press_keycode_with_flags_with_wrong_flags
-            assert_raises ArgumentError do
+            assert_raises ::Appium::Core::Error::ArgumentError do
               @driver.press_keycode 86, flags: 0x02
             end
           end
 
           # keypress
           def test_press_keycode_with_flags_with_wrong_metastate
-            assert_raises ArgumentError do
+            assert_raises ::Appium::Core::Error::ArgumentError do
               @driver.press_keycode 86, metastate: 0x02
             end
           end
@@ -121,14 +121,14 @@ class AppiumLibCoreTest
 
           # keypress
           def test_long_press_keycode_with_flags_with_wrong_flags
-            assert_raises ArgumentError do
+            assert_raises ::Appium::Core::Error::ArgumentError do
               @driver.long_press_keycode 86, flags: 0x02
             end
           end
 
           # keypress
           def test_long_press_keycode_with_flags_with_wrong_metastate
-            assert_raises ArgumentError do
+            assert_raises ::Appium::Core::Error::ArgumentError do
               @driver.long_press_keycode 86, metastate: 0x02
             end
           end